+++ test_runner/src/main/kotlin/ftl/gc/GcStorage.kt
@@ -80,11 +80,11 @@ object GcStorage {
         }
     }
 
-    fun uploadCiJUnitXml(testResult: JUnitTestResult, args: IArgs, fileName: String) {
-        if (args.resultsBucket.isBlank() || args.resultsDir.isBlank()) return
+    fun uploadReportResult(testResult: String, args: IArgs, fileName: String) {
+        if (args.resultsBucket.isBlank() || args.resultsDir.isBlank() || args.disableResultsUpload) return
         upload(
             filePath = fileName,
-            fileBytes = testResult.xmlToString().toByteArray(),
+            fileBytes = testResult.toByteArray(),
             rootGcsBucket = args.resultsBucket,
             runGcsPath = args.resultsDir
         )