+++ test_runner/src/main/kotlin/ftl/reports/CostReport.kt
@@ -2,6 +2,7 @@ package ftl.reports
 
 import ftl.args.IArgs
 import ftl.config.FtlConstants.indent
+import ftl.gc.GcStorage
 import ftl.json.MatrixMap
 import ftl.reports.util.IReport
 import ftl.reports.xml.model.JUnitTestResult
@@ -41,5 +42,6 @@ object CostReport : IReport {
         val output = generate(matrices)
         if (printToStdout) print(output)
         write(matrices, output, args)
+        GcStorage.uploadReportResult(output, args, fileName())
     }
 }