+++ test_runner/src/main/kotlin/ftl/reports/util/IReport.kt
@@ -2,7 +2,6 @@ package ftl.reports.util
 
 import ftl.args.IArgs
 import ftl.json.MatrixMap
-import ftl.reports.JUnitReport
 import ftl.reports.xml.model.JUnitTestResult
 import ftl.util.resolveLocalRunPath
 import ftl.util.write
@@ -14,7 +13,9 @@ interface IReport {
     fun reportName(): String {
         return this::class.java.simpleName
     }
+
     fun fileName() = reportName() + extension
+
     val extension: String
 
     fun reportPath(matrices: MatrixMap, args: IArgs): String {
@@ -22,7 +23,7 @@ interface IReport {
         return Paths.get(path, fileName()).toString()
     }
     fun write(matrices: MatrixMap, output: String, args: IArgs) {
-        val reportPath = JUnitReport.reportPath(matrices, args)
+        val reportPath = reportPath(matrices, args)
         reportPath.write(output)
     }
 }